+++ include/SW_Control.h
@@ -30,6 +30,7 @@ extern "C" {
 /* --------------------------------------------------- */
 void SW_CTL_init_ptrs(SW_ALL* sw) ;
 void SW_CTL_alloc_outptrs(SW_ALL* sw, LOG_INFO* LogInfo);
+void SW_ALL_deepCopy(SW_ALL* source, SW_ALL* dest, LOG_INFO* LogInfo);
 void SW_CTL_setup_domain(unsigned long userSUID,
                          SW_DOMAIN* SW_Domain,
                          LOG_INFO* LogInfo);
@@ -43,6 +44,7 @@ void SW_CTL_main(SW_ALL* sw, SW_OUTPUT_POINTERS* SW_OutputPtrs,
                  LOG_INFO* LogInfo); /* main controlling loop for SOILWAT  */
 void SW_CTL_RunSimSet(SW_ALL *sw_template, SW_OUTPUT_POINTERS SW_OutputPtrs[],
                       SW_DOMAIN *SW_Domain,
+                      SW_WALLTIME *SW_WallTime,
                       LOG_INFO *main_LogInfo);
 void SW_CTL_run_current_year(SW_ALL* sw, SW_OUTPUT_POINTERS* SW_OutputPtrs,
                              LOG_INFO* LogInfo);